+++ ChangeLog.md
@@ -3,6 +3,9 @@ ZIP File support for the XP Framework ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Added `add()` method which calls addFile() or addDir() depending on the
+  given type given.
+  (@thekid)
 * Added `in()` and `out()` methods as successors to the verbose forms 
   `getInputStream()` and `getOutputStream()`.
   (@thekid)